renaming js files for devMilestones and medMilestones models
